Refactor route names
